Extract MediaFetchSession and MediaSourceFetchResult to separate file
